# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Blog Documentation - OpenAPI 3.0",
    "description": "Dokumentasi API Blog",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
...
  "paths": {
...
    "/auth/authuser/{email}{password}": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "authenticate user",
        "description": "authUser",
        "operationId": "authUser",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "user email",
            "required": true,
            "schema": {
              "type": "string",
              "default": 1
            }
          },
          {
            "name": "password",
            "in": "path",
            "description": "user password",
            "required": true,
            "schema": {
              "type": "string",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/whoami"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Email sudah terdaftar",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Failed!"
                    },
                    "message": {
                      "type": "string",
                      "example": "you're not authorized!"
                    },
                    "data": {
                      "type": "string",
                      "example": null
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "authorization": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "whoami": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "Success!"
          },
          "message": {
            "type": "string",
            "example": "OK"
          },
          "data": {
            "type": "string",
            "example": {
              "user": {
                "data": "user data here"
              }
            }
          }
        }
      }
    },
...
    "securitySchemes": {
      "authorization": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header"
      }
    }
  }
}